Ensure that players only get applicable drops for the tool used
Previously it was possible to get some drops by simply breaking the block.
This commit changes some of that behaviour to ensure that a mojority of
blocks that require a tool to break (such as ores) will only drop their
items if the correct minimum tool is used.

This does not resolve all cases of bad drop mappings and may have some
edge cases regarding what requires a tool, however this is a starting
point for many of the future cases.