Improve biome decorations
Through the use of a revampped biome decorator structure, the following
decorations have been applied to new world generation.

- Tall grass (and ferns where applicable)
- Flowers (both double-tall and small)
- Sugar canes
- Dead bushes
- Pumpkins
- Melons
- Water lilies
- Cacti

The placement logic is currently not biome specific due to Glowstone not
having a true concept of a biome during world generation. Therefore,
currently the aforementioned decorations are applied to where ever they
fit as per the vanilla ruleset.