Implement resource pack event and methods for 1.8
This commit implements the event for the client reporting the status of
the application of a resource pack after it has been sent. This event may
be fired with illogical statuses due to bugs in the client's operations.

This also adds the applicable methods for plugins to set a specific
player's resource pack, or at least prompt them to, using the newly
implemented identifier system.